+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/Renderer.java
@@ -40,6 +40,7 @@ import java.lang.annotation.Documented;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
+import java.util.List;
 
 /**
  * Renders media read from a {@link SampleStream}.
@@ -86,8 +87,8 @@ public interface Renderer extends PlayerMessage.Target {
    * #MSG_SET_SCALING_MODE}, {@link #MSG_SET_CHANGE_FRAME_RATE_STRATEGY}, {@link
    * #MSG_SET_AUX_EFFECT_INFO}, {@link #MSG_SET_VIDEO_FRAME_METADATA_LISTENER}, {@link
    * #MSG_SET_CAMERA_MOTION_LISTENER}, {@link #MSG_SET_SKIP_SILENCE_ENABLED}, {@link
-   * #MSG_SET_AUDIO_SESSION_ID}, {@link #MSG_SET_WAKEUP_LISTENER} or {@link
-   * #MSG_SET_VIDEO_OUTPUT_RESOLUTION}. May also be an app-defined value (see {@link
+   * #MSG_SET_AUDIO_SESSION_ID}, {@link #MSG_SET_WAKEUP_LISTENER}, {@link #MSG_SET_VIDEO_EFFECTS} or
+   * {@link #MSG_SET_VIDEO_OUTPUT_RESOLUTION}. May also be an app-defined value (see {@link
    * #MSG_CUSTOM_BASE}).
    */
   @Documented
@@ -107,6 +108,7 @@ public interface Renderer extends PlayerMessage.Target {
         MSG_SET_SKIP_SILENCE_ENABLED,
         MSG_SET_AUDIO_SESSION_ID,
         MSG_SET_WAKEUP_LISTENER,
+        MSG_SET_VIDEO_EFFECTS,
         MSG_SET_VIDEO_OUTPUT_RESOLUTION
       })
   public @interface MessageType {}
@@ -210,11 +212,16 @@ public interface Renderer extends PlayerMessage.Target {
    */
   int MSG_SET_PREFERRED_AUDIO_DEVICE = 12;
   /**
+   * The type of a message that can be passed to a video renderer. The message payload should be a
+   * {@link List} containing {@linkplain Effect video effects}.
+   */
+  int MSG_SET_VIDEO_EFFECTS = 13;
+  /**
    * The type of a message that can be passed to a video renderer to set the desired output
    * resolution. The message payload should be a {@link Size} of the desired output width and
    * height. Use this method only when playing with video {@linkplain Effect effects}.
    */
-  int MSG_SET_VIDEO_OUTPUT_RESOLUTION = 13;
+  int MSG_SET_VIDEO_OUTPUT_RESOLUTION = 14;
   /**
    * Applications or extensions may define custom {@code MSG_*} constants that can be passed to
    * renderers. These custom constants must be greater than or equal to this value.